+++ tests/unit_tests/openvpn/test_cryptoapi.c
@@ -467,7 +467,7 @@ test_parse_hexstring(void **state)
     unsigned char hash[255];
     (void)state;
 
-    for (int i = 0; i < _countof(valid_str); i++)
+    for (size_t i = 0; i < _countof(valid_str); i++)
     {
         int len = parse_hexstring(valid_str[i], hash, _countof(hash));
         assert_int_equal(len, sizeof(test_hash));
@@ -475,7 +475,7 @@ test_parse_hexstring(void **state)
         memset(hash, 0, _countof(hash));
     }
 
-    for (int i = 0; i < _countof(invalid_str); i++)
+    for (size_t i = 0; i < _countof(invalid_str); i++)
     {
         int len = parse_hexstring(invalid_str[i], hash, _countof(hash));
         assert_int_equal(len, 0);
